backend/pyproject.toml
[tool.poetry]
name = "aidaytrading-backend"
version = "0.1.0"
description = "Aidaytrading FastAPI backend foundation"
authors = ["Aidaytrading Team <dev@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.0"
uvicorn = { version = "0.29.0", extras = ["standard"] }
sqlalchemy = "2.0.29"
aiosqlite = "0.19.0"
alembic = "1.13.1"
passlib = { version = "1.7.4", extras = ["bcrypt"] }
bcrypt = "3.2.2"
python-dotenv = "1.0.1"
loguru = "0.7.2"
pydantic-settings = "2.2.1"
python-jose = { version = "3.3.0", extras = ["cryptography"] }
email-validator = "2.1.1"
jinja2 = "3.1.3"
anyio = "4.3.0"
apscheduler = "3.10.4"
firebase-admin = "6.5.0"
pytz = "2024.1"

[tool.poetry.group.dev.dependencies]
black = "24.3.0"
flake8 = "7.0.0"
isort = "5.13.2"
pytest = "8.1.1"
pytest-asyncio = "0.23.6"
pytest-cov = "4.1.0"
pytest-timeout = "2.3.1"
httpx = "0.27.0"
mypy = "1.9.0"
pre-commit = "3.6.2"
coverage = "7.4.4"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\\.(py|pyi)$'
exclude = '''
/(\.git|\.venv|build|dist|reports|__pycache__)/
'''

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
multi_line_output = 3

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]
namespace_packages = true

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
